backend/requirements.txt

Django==4.1.1
django-cors-headers==3.13.0
django-crum==0.7.9
djangorestframework==3.14.0
selenium==4.5.0
psycopg2

